requirements.txt
accelerate==0.29.3
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
auto-gptq==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
dataclasses-json==0.6.4
datasets==2.18.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.2.0
gekko==1.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.38
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.38
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.16.0
optimum==1.19.0
orjson==3.10.0
packaging==23.2
pandas==2.0.3
peft==0.10.0
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic-core==2.16.3
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rouge==1.0.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.6.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
triton==2.2.0
types-requests==2.31.0.20240402
typing-extensions==4.10.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
